requirements.txt
django
djangorestframework
django-model-utils
psycopg2
